Move resolved obj IDs into std db package
